requirements.txt
ftfy
regex
tqdm
omegaconf
pytorch-lightning
kornia
einops
wget
stegano
python-xmp-toolkit
imgtag
pillow==7.1.2 
black
flake8